requirements.txt
streamlit
google-generativeai
requests
streamlit-mic-recorder
